Refactoring: Create and use OpenMetricsBaseEntity class
