+++ custom_components/openmetrics/update.py
@@ -12,18 +12,12 @@ from homeassistant.const import (
     EntityCategory,
 )
 from homeassistant.core import HomeAssistant
-from homeassistant.helpers.device_registry import DeviceEntryType, DeviceInfo
+from homeassistant.helpers.device_registry import DeviceInfo
 from homeassistant.helpers.entity_platform import AddEntitiesCallback
-from homeassistant.helpers.update_coordinator import (
-    CoordinatorEntity,
-    DataUpdateCoordinator,
-)
+from homeassistant.helpers.update_coordinator import DataUpdateCoordinator
 
-from .const import (
-    DOMAIN,
-    RESOURCE_TYPE_CONTAINER,
-    RESOURCE_TYPE_NODE,
-)
+from .const import DOMAIN
+from .entity import OpenMetricsBaseEntity, async_setup_entities, create_device_info
 from .metrics.data import ResourceInfoData
 from .providers.node_exporter import (
     METRIC_NODE_OS_UPDATE_INFO,
@@ -44,86 +38,45 @@ UPDATE_ENTITIES: dict[str, UpdateEntityDescription] = {
 async def async_setup_entry(
     hass: HomeAssistant, entry: ConfigEntry, async_add_entities: AddEntitiesCallback
 ) -> None:
-    """Set up OpenMetrics sensors based on a config entry."""
-    # Get coordinator
-    coordinator = hass.data[DOMAIN][entry.entry_id]["coordinator"]
-    host = hass.data[DOMAIN][entry.entry_id]["host"]
-    metric_keys = entry.data["metrics"]
-    # Create sensors
-    for resource in coordinator.resources.values():
-        # Create entities for each resource
-        entities = create_resource_entities(resource, host, coordinator, metric_keys)
-        # Add entities to hass
-        async_add_entities(entities)
+    """Set up OpenMetrics update entities based on a config entry."""
+    await async_setup_entities(
+        hass, entry, async_add_entities, create_resource_update_entities
+    )
 
 
-def create_resource_entities(
+def create_resource_update_entities(
     resource: ResourceInfoData,
     host: str,
-    coordinator,
+    coordinator: DataUpdateCoordinator,
     metric_keys: list[str],
 ) -> list[Any]:
-    """Create entities for the given resource."""
+    """Create update entities for the given resource."""
     entities = []
     unique_id = f"{host}_{resource.name}"
     if resource.is_virtual:
         via_device = (DOMAIN, f"{host}_{resource.via_resource}")
         device_info = create_device_info(unique_id, resource, via_device)
+        update_entity_descriptions = {}
     else:
         device_info = create_device_info(unique_id, resource)
+        update_entity_descriptions = UPDATE_ENTITIES
+
     # Create update entities
-    for key, description in UPDATE_ENTITIES.items():
+    for key, description in update_entity_descriptions.items():
         if key in metric_keys:
             entity = OpenMetricsUpdateEntity(
                 coordinator, description, device_info, resource
             )
             entities.append(entity)
-    # Return created entities
+
     return entities
 
 
-def create_device_info(
-    unique_id: str,
-    resource: ResourceInfoData,
-    via_device: tuple[str, str] | None = None,
-) -> DeviceInfo:
-    """Create a device info object for a resource."""
-    # Create device info object with the minimal required fields (-> Generic provider)
-    device_info = DeviceInfo(
-        name=resource.name,
-        identifiers={(DOMAIN, unique_id)},
-    )
-    # Set model if provided
-    if resource.model:
-        device_info["model"] = resource.model
-    # Set serial number if provided
-    if resource.serial_number:
-        device_info["serial_number"] = resource.serial_number
-    # Set via_device if provided
-    if via_device:
-        device_info["via_device"] = via_device
-    # Set resource type related attributes
-    if resource.type == RESOURCE_TYPE_CONTAINER:
-        # Set software version
-        device_info["sw_version"] = resource.version
-        # Set entry type
-        device_info["entry_type"] = DeviceEntryType.SERVICE
-    elif resource.type == RESOURCE_TYPE_NODE:
-        if resource.software and resource.version:
-            device_info["sw_version"] = f"{resource.software} {resource.version}"
-        else:
-            device_info["sw_version"] = resource.software
-    # Return the device info object
-    return device_info
-
-
-class OpenMetricsUpdateEntity(CoordinatorEntity, UpdateEntity):
+class OpenMetricsUpdateEntity(OpenMetricsBaseEntity, UpdateEntity):
     """Representation of an OpenMetrics update entity."""
 
-    _attr_has_entity_name = True
     _attr_device_class = UpdateDeviceClass.FIRMWARE
     entity_description: UpdateEntityDescription
-    device_info: DeviceInfo
 
     def __init__(
         self,
@@ -132,25 +85,16 @@ class OpenMetricsUpdateEntity(CoordinatorEntity, UpdateEntity):
         device_info: DeviceInfo,
         resource: ResourceInfoData,
     ) -> None:
-        """Initialize the sensor."""
-        super().__init__(coordinator)
+        """Initialize the update entity."""
         self.entity_description = description
-        self.device_info = device_info
-        identifier = next(iter(device_info.get("identifiers", {})))
-        self._attr_unique_id = f"{identifier[1]}_{description.key}"
-        self.entity_id = f"sensor.{self._attr_unique_id.replace('.', '_').replace(':', '_').replace('/', '_')}"
-        self._attr_title = resource.software
         self.resource = resource
+        super().__init__(coordinator, device_info)
 
-    @property
-    def translation_key(self) -> str | None:
-        """Return the translation key to translate the entity's name and states."""
-        return self.entity_description.translation_key
+        self._attr_title = resource.software
 
-    @property
-    def unique_id(self) -> str | None:
-        """Return the unique ID."""
-        return self._attr_unique_id
+        if self._attr_unique_id:
+            # Replace invalid characters in the unique ID with underscores
+            self.entity_id = f"update.{self._attr_unique_id.replace('.', '_').replace(':', '_').replace('/', '_')}"
 
     @property
     def installed_version(self) -> str | None:
